Remove unused fixture imports
